Check file existence in periscope_argus_add_file.  Don't accept input from stdin.
Argus allows input from standard in, we won't be using that feature.  

Also, Argus delays checking whether files exist until when it needs to process it -
we may need to check earlier than that, so return error in periscope_argus_add_file
if access(pathname, F_OK) fails and don't add the pathname to the Argus input list.