Modify flow semantics to take account of time values with microseconds better.
My hackshift method was to add a decimal percentage corresponding to the usec parameter
to the seconds parameter, but this is inaccurate even when forcing a double-float type.
Use usecs separately, and add methods START-TIME and END-TIME to properly calculate the
start and end time of a flow, or a particular source: 

(start-time flow)
(start-time (source flow))
(end-time (dest flow))
etc.