Implement a much-improved DNS reverse lookup architecture.
DNS requests are pooled into a queue, where they are processed asynchronously in another
thread.  The interface to the user is the function HOSTNAME, which returns two values -
the canonical hostname of the supplied IP address -  if it exists and has been determined -
and a keyword indicating the status of the lookup, which is one of :FOUND, :NXDOMAIN, or
:PROCESSING.  The application can do what it likes with these identifiers.  If the hostname
has not or cannot be determined, the result of the function IP-STRING applied to the
input parameter is returned along with its corresponding status (either :NXDOMAIN or :PROCESSING).