Make service reports available hourly.
Also removes handling of FILTERs from PERIODIC-REPORT, since it is handled
in the hourly report handler now, as well as the time of the report.

The new scheme for displaying hourly reports allows for any number of reports
to be displayed simultaneously, as MAKE-FILTERED-REPORTS generates both
SERVICE and PERIODIC-REPORT reports on the filtered lists.