Lots of fixes to user login handling.
Calling (USER) will now remove the current session if the user it points to
does NOT have the session recorded in their SESSION slot; this will prevent
being unable to login in most cases.

The /login page should also NOT shadow *WEB-LOGIN-REQUIRED-P*, but rather
VALID-SESSION-OR-LOSE should make an exception explicitly for /login.

Finally, if /do-login fails and (LOGIN-REQUIRED-P) is T, redirect directly
back to /login.