Quick hack to fix SELECT-DAO queries.
We can't insert integer represented inet fields, but that's what we get back from PostgreSQL
with my SQL reader, so we use a Lisp->SQL string converter for FLOW-HOST.  SETF the HOST-IP
of a SELECTed host entry to a blank FLOW-HOST with only HOST-IP filled in.

This is a hack for now; a better solution must be implemented (which will also slightly
improve performance for cache misses, i think).