Require *COLLECTOR-LOOKUP-TABLE* to be value-weak.
The collector-pointer lookup table maps libperiscope collector pointers to
lisp COLLECTOR objects.  When the COLLECTOR objects are finalized (established
with TG:FINALIZE) we remove the entry from the hash table; the problem is, the
finalizer is never called because a reference still exists in the hash table.

Enter weak hash tables.  Since the table is now value-weak, whenever the value's
only reference is the hash table, it is allowed to be removed and garbage collected.
Huzzah